+++ frontend/src/app/pages/exercises/exercise/shared/didactic-overview/didactic-overview-modal/open-didactic-overview-modal.ts
@@ -1,5 +1,5 @@
 import type { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { DidacticOverviewModalComponent } from '../didactic-overview-modal/didactic-overview-modal.component';
+import { DidacticOverviewModalComponent } from './didactic-overview-modal.component';
 
 export function openDidacticOverviewModal(ngbModalService: NgbModal) {
     ngbModalService.open(DidacticOverviewModalComponent, { size: 'xl' });